python/sampleONNXResNet/requirements.txt
# pip install -r requirements.txt

numpy>=1.18.5
pillow>=7.1.2
